adding custom hook
